requirements.dev.txt
flake8>=3.9.2,<3.10
django-cors-headers>=4.3.1,<4.4
